+++ frontend/src/features/navigation/header/authButtons.tsx
@@ -1,0 +1,12 @@
+import Button from '@mui/material/Button';
+
+const AuthButtons = () => (
+  <>
+    <Button color="inherit">Log in</Button>
+    <Button variant="contained" color="primary">
+      Sign up
+    </Button>
+  </>
+);
+
+export default AuthButtons;
